Rename new_tests to tests_new to fix workflows
