+++ src/qibo/tests_new/test_abstract_circuit.py
