+++ src/qibo/tests_new/test_abstract_circuit_qasm.py
@@ -1,7 +1,7 @@
 import pytest
 from qibo import __version__
 from qibo.abstractions import gates
-from qibo.new_tests.test_abstract_circuit import Circuit
+from qibo.tests_new.test_abstract_circuit import Circuit
 
 
 def assert_strings_equal(a, b):
