+++ src/qibo/tests_new/test_backends_init.py
