+++ src/qibo/tests_new/test_core_circuit.py
